requirements.txt
Django==3.1
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
Pillow==8.4.0
pandas